feat(Reports): :zap: Implementation of reports in HTML and JSON format
The jest-html-reports library is used to generate the HTML reports. It is the most up-to-date library with optimal graphical consistency, allowing for the identification of each factor in the different test cases. A JsonReporter class is also implemented for creating custom JSON reports. Finally, the Jest configuration is modified to centralize all reports.